raspberry_docker/requirements.txt
requests
configparser
uuid
pandas
pillow